add controller for auth routes
